backend/requirements.txt
alembic==1.0.5
amqp==2.3.2
aniso8601==4.0.1
asgiref==3.4.1
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.3
blessed==1.15.0
blinker==1.4
Bootstrap-Flask==2.2.0
boto3==1.7.44
botocore==1.10.44
cached-property==1.4.3
cement==2.8.2
certifi==2018.4.16
cffi==1.15.1
charset-normalizer==2.0.12
Click==7.0
colorama==0.3.9
coverage==4.5.1
cryptography==3.2
csvalidate==1.1.1
Django==3.2.16
docopt==0.6.2
docutils==0.14
dominate==2.3.5
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Cors==3.0.10
Flask-CSV==1.2.0
Flask-Excel==0.0.7
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Menu==0.7.0
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.1.5
Flask-WTF==0.14.2
gunicorn==20.1.0
hurry==1.1
hurry.filesize==0.9
idna==2.6
image==1.5.33
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.11.3
jmespath==0.9.3
jsonschema==2.6.0
kombu==4.2.1
lml==0.0.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==3.2.2
numpy==1.19.5
opencv-python==4.6.0.66
openpyxl==2.5.4
paramiko==2.6.0
passlib==1.7.1
pathspec==0.5.5
Pillow==8.2.0
psycopg2-binary==2.9.3
py==1.10.0
pycparser==2.19
pycryptodome==3.6.6
pyexcel==0.5.8
pyexcel-io==0.5.7
pyexcel-webio==0.1.4
pyexcel-xls==0.5.7
pyexcel-xlsx==0.5.6
PyJWT==1.6.4
PyNaCl==1.3.0
python-dateutil==2.7.5
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.7
PyYAML==5.4
requests==2.27.1
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.3.7
sqlparse==0.4.3
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
typing_extensions==4.1.1
urllib3==1.26.5
vine==1.1.4
visitor==0.1.3
websocket-client==0.48.0
Werkzeug==0.15.5
WTForms==2.2.1
xlrd==1.1.0
xlwt==1.3.0
